reduce components to pages for now
