+++ frontend/new/pages/index.tsx
@@ -1,17 +1,33 @@
-import Head from 'next/head'
+import { useRouter } from 'next/router'
 
-import Home from '@/components/Home/Home'
+import api from '@/lib/api'
+import css from './index.module.css'
 
 export default function Index() {
-  return (
-    <>
-      <Head>
-        <title>treclo</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Home />
-    </>
-  )
+    const router = useRouter()
+
+    function createWorkspace() {
+        api
+            .post('workspaces')
+            .then(({ identifier}) => router.push(`workspaces/${identifier}`))
+    }
+
+    return (
+        <div className={css.home}>
+            <h2 className={css.welcome}>
+                <span className={css.hello}> hello </span>
+                from Kanbad!
+            </h2>
+
+            <p className={css.caption}>
+                Workspaces give you places to make things like boards and cards... and magic!
+                Create a new one or, if you're really lucky, get a friend to share a workspace with you.
+            </p>
+            <div className={css.create}>
+                <button className={css.button} onClick={createWorkspace}>
+                    Create workspace
+                </button>
+            </div>
+        </div>
+    )
 }
